Added new properties and minor fixes
The "avatarMode" property has been added. ['true' or true or 1].
When the property is enabled, the list has a multi-line view with an avatar.
http://materializecss.com/collections.html#circle

Added property "selectedMode".  ['true' or true or 1].
In this mode, list items can be selected.
The result 'id' of the selected element is placed in the synchronized property 'selectedId' (: selected-id.sync = "your_variable")
or the event is caught: @update: selectedId = your_function